+++ boot/i386/nasm/ndisasm.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -218,7 +219,7 @@ int main(int argc, char **argv) {
 	if (lenread == 0)
 	    eof = TRUE;		       /* help along systems with bad feof */
 	p += lenread;
-	if (offset == nextsync) {
+	if (offset == (long)nextsync) {
 	    if (synclen) {
 		printf("%08lX  skipping 0x%lX bytes\n", offset, synclen);
 		offset += synclen;
@@ -230,7 +231,7 @@ int main(int argc, char **argv) {
 	while (p > q && (p - q >= INSN_MAX || lenread == 0)) {
 	    lendis = disasm (q, outbuf, bits, offset, autosync);
 	    if (!lendis || lendis > (p - q) ||
-		lendis > nextsync-offset)
+		lendis > (int)(nextsync-offset))
 		lendis = eatbyte (q, outbuf);
 	    output_ins (offset, q, lendis, outbuf);
 	    q += lendis;
