+++ cctools/file/is_tar.c
@@ -3,8 +3,6 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -65,7 +63,7 @@ int nbytes;
 	register int	sum, recsum;
 	register char	*p;
 
-	if (nbytes < sizeof(union record))
+	if (nbytes < (int)sizeof(union record))
 		return 0;
 
 	recsum = from_oct(8,  header->header.chksum);
