+++ cctools/file/readfat.c
@@ -3,8 +3,6 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -55,7 +53,7 @@ int nbytes)
     unsigned char tmpbuf[HOWMANY+1];	/* one extra for terminating '\0' */
     
 
-	if(nbytes < sizeof(struct fat_header)){
+	if(nbytes < (int)sizeof(struct fat_header)){
 	    return;
 	}
 	memcpy(&fat_header, buf, sizeof(struct fat_header));
@@ -63,7 +61,7 @@ int nbytes)
 	swap_fat_header(&fat_header, NX_LittleEndian);
 #endif /* __LITTLE_ENDIAN__ */
 	arch_size = fat_header.nfat_arch * sizeof(struct fat_arch);
-	if(arch_size + sizeof(struct fat_header) > nbytes){
+	if(arch_size + sizeof(struct fat_header) > (unsigned long)nbytes){
 	    return;
 	}
 	arch_buf = malloc(nbytes);
@@ -83,7 +81,7 @@ int nbytes)
 	    /*
 	     * try looking at the first HOWMANY bytes
 	     */
-	    if ((tbytes = read(fd, (char *)tmpbuf, HOWMANY)) == -1) {
+	    if ((int)(tbytes = read(fd, (char *)tmpbuf, HOWMANY)) == -1) {
 		error("read failed (%s).\n", strerror(errno));
 		/*NOTREACHED*/
 	    }
