+++ cctools/misc/size.c
@@ -3,8 +3,6 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -92,7 +90,7 @@ int argc,
 char **argv,
 char **envp)
 {
-    unsigned long i;
+    int i;
     enum bool args_left;
     struct flags flag;
     struct arch_flag *arch_flags;
