+++ cctools/otool/m68k_disasm.c
@@ -3,8 +3,6 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -1945,9 +1943,10 @@ enum bool verbose)
 				       sect + length,
 				       fpsize[(specop1 >> 10) & 0x7]);
 		    if(((specop1 >> 10) & 0x7) == 0x3){
-			printf("{#%u}\n", specop1 & 0x0040 ?
-			       (specop1 & 0x7f) | 0xffffff80 :
-				specop1 & 0x7f);
+			printf("{#%u}\n", (unsigned int)(specop1 & 0x0040) ?
+			       (unsigned int)(specop1 & 0x7f) |
+				(unsigned int)0xffffff80 :
+				(unsigned int)(specop1 & 0x7f));
 		    } else if(((specop1 >> 10) & 0x7) == 0x7){
 			printf("{%s}\n", dregs[(specop1 >> 4) & 0x7]);
 		    } else
